requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic>=0.40.0
duckduckgo-search==6.3.7
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
